requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.15
pydantic-settings>=2.8
python-dotenv>=1.0
PyYAML>=6.0
pytest>=8.3
pytest-asyncio>=0.25
httpx>=0.28
dingtalk-stream==0.24.3
openai>=2.0
